+++ package.json
@@ -24,6 +24,7 @@
     "clean-webpack-plugin": "^2.0.2",
     "compression-webpack-plugin": "^3.0.0",
     "css-loader": "^2.1.1",
+    "csv-writer": "^1.5.0",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.15.1",
     "eslint": "^5.16.0",
@@ -58,6 +59,7 @@
     "express-static-gzip": "^2.0.5",
     "moment": "^2.24.0",
     "node-cache": "^4.2.1",
+    "pg": "^7.17.1",
     "react": "^16.12.0",
     "react-apexcharts": "^1.3.3",
     "react-dom": "^16.12.0",
