+++ server/messages.js
@@ -1,0 +1,12 @@
+module.exports = {
+  fetches: {
+    twBearer: 'Failed to fetch Twitter bearer token',
+    twTimeline: 'Failed to fetch latest tweets',
+  },
+  uploads: {
+    esError: 'Failed to upload new tweets to elastic search',
+    esSuccess: 'Successfully uploaded new tweets to elastic search',
+    pgError: 'Failed to upload new tweets to the database',
+    pgSuccess: 'Successfully uploaded new tweets to the database',
+  },
+};
